:recycle: Move filter related files to a specific folder
