+++ src/components/transactionsV2/filters/filterBar.css
@@ -1,4 +1,4 @@
-@import '../app/variablesV2.css';
+@import '../../app/variablesV2.css';
 
 .container {
   width: 100%;
